vici: fix validation of whether virtual IPs are within IP pool
The function have_vips_from_pool first calculates the start and end addresses of the IP pool. During the verification of whether a virtual IP is within these boundaries, the end address is now considered inclusive rather than exclusive.

This function is exclusively used in the create_attribute_enumerator interface of the attribute provider in the vici plugin, which is called during IKE_AUTH and delivers the attributes for the configured configuration payload.