wip: testing: Use different version of TKM on trixie
The API for libgmpada has change with 1.6 in a way that's not
backwards-compatible.  So we use a different revision that includes
the required changes depending on the Debian version.

wip: update rev, the xfrm-ada change is currenty necessary because of
another commit in the tkm branch (to implement esa_select())