wip: child-create: Add support for KE negotiation during IKE_AUTH
We prevent handling any KE payloads, just in case a peer incorrectly
sends us one.

wip: it might be a bit confusing that a proposal with KE methods is logged
and shown in the status output while the key is not actually based on them.
maybe documenting this thoroughly is enough. but perhaps we need a flag
to mark the sa (like we did for the initial simplified rekey approach)
so we could at least change the status display...(question would be how
that's conveyed e.g. via vici and what swanctl would do with it)